+++ app/entities/order/schemas/schema_order.py
@@ -6,7 +6,7 @@ from entities.order.models_order import OrderModel
 from entities.service_name.models_service_name import ServiceNameModel
 from entities.users.models_user import ClientModel, EmployeeModel
 from models.choices import StatusOrder
-from routers.choices import OrderFilter
+from entities.order.choices import OrderFilter
 from schemas.base_schema import PaginationSchema, BasePydanticSchema
 from entities.category.schemas_category import CategoryInSchema
 from core.utils.pagination import Pagination
