+++ app/tests/test_order.py
@@ -13,7 +13,7 @@ from entities.order.models_order import OrderModel
 from tests.utils import url_reverse
 from models.choices import StatusOrder
 from tests.conftest import MOCK_FREEZE_TIME, engine
-from routers.choices import OrderFilter
+from entities.order.choices import OrderFilter
 from entities.order.models_order import OrderInSchema
 from services.stub_init_service import T_BOOK_DATE
